src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Murat Kamaletdinov <kamaletdinov2425@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
psycopg2-binary = "^2.8.6"
drf-yasg = "^1.20.0"
pytest = "^6.2.4"
djangorestframework = "^3.12.4"
factory-boy = "^3.2.0"
pytest-django = "^4.3.0"
coverage = "^5.5"

[tool.poetry.dev-dependencies]
black = "^21.5b1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
